requirements/common.txt
# Core
pytz==2020.1
Django==3.0.8
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.14.1.144

# For the persistence stores
psycopg2-binary==2.8.5
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.11.0
Markdown==3.2.2
django-filter==2.3.0
drf-yasg

# Developer Tools
ipdb==0.13.3
ipython==7.16.1
mkdocs==1.1.2
flake8==3.8.3
isort==5.2.2
